backend/requirements.txt
# Flask Backend Dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
python-dotenv==1.0.0
gunicorn==21.2.0

# Database drivers (PyMySQL is pure Python - no compilation needed)
PyMySQL==1.1.0
cryptography==41.0.0

# Utilities
email-validator==2.1.0
